Properly support binding key releases
Now, the keyboard keeps track of a single array of keys pressed, similar
to th pointer, and the client handler keeps track of an array of keys to
be sent on enter.

This way, we can keep track of the binding used when a key is invoked,
in order to ensure that the same binding is invoked when the key is
released.