keyboard: Ignore unpaired key release events
If we see a release for a key we didn't know was pressed, it is because
the key was pressed when we opened the device. These are safe to ignore
because our internal state of the keyboard considered them unpressed.

This fixes a crash when the swc VT is switched to from a text VT, due to
accidently handling a key release like a key press.