window: Use button given in handler when calling the original handler
When a button is released normalling during a client-initiated resize or
move, we can't (and shouldn't because we already have the button passed
as an argument) look up the button by serial, because this changes to
the serial used for the release event.

Instead, add an extra button parameter to end_interaction, and only look
up the button by serial when it is NULL (when we end the interactions
prematurely).

This fixes a crash for client-initiated moves/resizes.