seat: Listen for uevents with group mask 4 by default
This way, the device manager can rebroadcast the uevents from the
kernel *after* setting the permissions.

Since Linux 5.8 no longer requires root to drop DRM master, we no
longer require setuid for the launcher. However, this means we can
longer expect to be able to open input devices based on uevents
from the kernel. Instead, expect the device manager to rebroadcast
the uevents when it has handled them.

Users that want to restore the previous behavior can build with
SWC_NETLINK_MASK=1.